requirements.txt

streamlit==1.35.0
opencv-python-headless==4.9.0.80
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.3.0
